requirements.txt
python-dotenv==1.0.1
langchain==0.2.0
langchain-community==0.2.0
langchain-experimental==0.0.59 
streamlit==1.34.0 
openai==1.30.1
tabulate==0.9.0